app/requirements.txt
icecream
bs4
selenium
webdriver-manager
psycopg2-binary
fastapi
pydantic
uvicorn
